+++ src/pages/MyPage.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import Purchase from '../Purchase';
-import UserProfile from '../../components/user/UserProfile';
-import MyMenu from '../../components/user/MyMenu';
+import Purchase from './Purchase';
+import UserProfile from '../components/user/UserProfile';
+import MyMenu from '../components/user/MyMenu';
 import styled from 'styled-components';
-import { getCookie } from '../../utils/cookie';
-import AlertLoginState from '../../components/common/AlertLoginState';
+import { getCookie } from '../utils/cookie';
+import AlertLoginState from '../components/common/AlertLoginState';
 
 const MyPage = () => {
   return (
